refactored mainactivity into several files, fixed bug that forced the user to restart the app the first time it was launched
