+++ app/cpp/CMakeLists.txt
@@ -5,7 +5,7 @@ project("testcpp")
 set(OpenCV_STATIC ON)
 
 
-set(OpenCV_DIR C:/Users/jcull/Downloads/opencv-4.9.0-android-sdk/OpenCV-android-sdk/sdk/native/jni)
+set(OpenCV_DIR C:/Users/Admin/Desktop/OpenCV-android-sdk/sdk/native/jni)
