+++ app/cpp/OpenCV_Func.cpp
@@ -455,7 +455,7 @@ Java_com_example_blodpool_LiveCamera_getarea(JNIEnv *env, jobject thiz) {
 }
 
 extern "C" JNIEXPORT void JNICALL
-Java_com_example_blodpool_MainActivity_findobjectinfo(JNIEnv *env, jobject thiz, jlong mat_addy, jint x_addy, jint y_addy) {
+Java_com_example_blodpool_ClickReferenceActivity_findobjectinfo(JNIEnv *env, jobject thiz, jlong mat_addy, jint x_addy, jint y_addy) {
     int x = static_cast<int>(x_addy);
     int y = static_cast<int>(y_addy);
     cv::Mat &mat = *(cv::Mat*) mat_addy;
@@ -467,7 +467,7 @@ Java_com_example_blodpool_MainActivity_findobjectinfo(JNIEnv *env, jobject thiz,
 }
 
 extern "C" JNIEXPORT void JNICALL
-Java_com_example_blodpool_MainActivity_centerobjectinfo(JNIEnv *env, jobject thiz, jlong mat_addy) {
+Java_com_example_blodpool_ClickReferenceActivity_centerobjectinfo(JNIEnv *env, jobject thiz, jlong mat_addy) {
     cv::Mat &mat = *(cv::Mat*) mat_addy;
 
     //cv::rotate(mat, mat, cv::ROTATE_90_CLOCKWISE);
@@ -478,7 +478,7 @@ Java_com_example_blodpool_MainActivity_centerobjectinfo(JNIEnv *env, jobject thi
 
 
 extern "C" JNIEXPORT void JNICALL
-Java_com_example_blodpool_MainActivity_getimage(JNIEnv *env, jobject thiz, jlong mat_addy) {
+Java_com_example_blodpool_ClickReferenceActivity_getimage(JNIEnv *env, jobject thiz, jlong mat_addy) {
     cv::Mat &mat = *(cv::Mat*) mat_addy;
 
     //cv::rotate(mat, mat, cv::ROTATE_90_CLOCKWISE);
@@ -490,7 +490,7 @@ Java_com_example_blodpool_MainActivity_getimage(JNIEnv *env, jobject thiz, jlong
 
 
 extern "C" JNIEXPORT float JNICALL
-Java_com_example_blodpool_MainActivity_getarea(JNIEnv *env, jobject thiz) {
+Java_com_example_blodpool_ClickReferenceActivity_getarea(JNIEnv *env, jobject thiz) {
 
     return getArea();
 }
