+++ app/src/main/java/com/example/blodpool/SettingsActivity.kt
@@ -1,0 +1,238 @@
+package com.example.blodpool
+
+import android.annotation.SuppressLint
+import android.app.AlertDialog
+import android.content.Intent
+import android.net.Uri
+import android.os.Bundle
+import android.os.Environment
+import android.widget.Button
+import android.widget.EditText
+import android.widget.ImageButton
+import android.widget.LinearLayout
+import android.widget.TextView
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+
+class SettingsActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_tutorial) // Set the layout file here
+
+        displaySettingsPage()
+    }
+
+    @SuppressLint("SetTextI18n")
+    fun chooseUnit(){
+        setContentView(R.layout.choose_unit)
+        val dlbutton = findViewById<Button>(R.id.button4)
+        val flozbutton = findViewById<Button>(R.id.button8)
+        val backbutton = findViewById<ImageButton>(R.id.buttonbaackk)
+        val unitused = findViewById<TextView>(R.id.textView3)
+
+        unitused.text = "current unit: $unittobedisplayed"
+
+
+        backbutton.setOnClickListener{
+            displaySettingsPage()
+        }
+
+        dlbutton.setOnClickListener{
+            unitcalc = 1f
+            unittobedisplayed = "dl"
+            chooseUnit()
+        }
+
+        flozbutton.setOnClickListener{
+            unitcalc = 3.38140227f
+            unittobedisplayed = "fl.oz"
+            chooseUnit()
+        }
+
+    }
+
+    fun displaySettingsPage(){
+        setContentView(R.layout.settings)
+        val aboutUsButton = findViewById<ImageButton>(R.id.aboutus)
+        val displayLiquidsButton = findViewById<ImageButton>(R.id.liquid)
+        val goBackToLandingPageButton = findViewById<ImageButton>(R.id.backbtn123)
+        val goToChooseUnitPageButton = findViewById<ImageButton>(R.id.language)
+
+        goToChooseUnitPageButton.setOnClickListener{
+            chooseUnit()
+        }
+
+        displayLiquidsButton.setOnClickListener{
+            displayCustomLiquids()
+        }
+
+        aboutUsButton.setOnClickListener{
+            displayAboutUsPage()
+        }
+
+        goBackToLandingPageButton.setOnClickListener{
+            finish()
+        }
+
+    }
+
+    private fun customliquids() {
+
+        val nameInputDialog = AlertDialog.Builder(this)
+        val nameInputEditText = EditText(this)
+        nameInputDialog.setTitle("Insert Liquid Name")
+        nameInputDialog.setView(nameInputEditText)
+        nameInputDialog.setPositiveButton("Confirm") { dialog, _ ->
+            val nameInput = nameInputEditText.text.toString()
+            if(nameInput != null){
+
+
+                val densityInputDialog = AlertDialog.Builder(this)
+                val densityInputEditText = EditText(this)
+                densityInputDialog.setTitle("Insert Density in kg/m^3")
+                densityInputDialog.setView(densityInputEditText)
+                densityInputDialog.setPositiveButton("Confirm") { dialog, _ ->
+                    val densityInputString = densityInputEditText.text.toString()
+                    try {
+                        val densityInput = densityInputString.toFloat()
+                        if (densityInput != null) {
+                            // Density input is valid, proceed to surface tension input
+                            dialog.dismiss()
+
+                            // Show a dialog to input surface tension
+                            val surfaceTensionInputDialog = AlertDialog.Builder(this)
+                            val surfaceTensionInputEditText = EditText(this)
+                            surfaceTensionInputDialog.setTitle("Insert Surface Tension in N/m")
+                            surfaceTensionInputDialog.setView(surfaceTensionInputEditText)
+                            surfaceTensionInputDialog.setPositiveButton("Confirm") { _, _ ->
+                                val surfaceTensionInput = surfaceTensionInputEditText.text.toString()
+                                try {
+                                    val surfaceTensionFloat = surfaceTensionInput.toFloat()
+                                    val surfaceTension = surfaceTensionFloat
+
+
+                                    // Valid surface tension input
+                                    val liquidManager = LiquidManager()
+                                    liquidManager.saveLiquid(nameInput, densityInput, surfaceTension, getExternalFilesDir(Environment.DIRECTORY_PICTURES))
+                                    displayCustomLiquids()
+                                } catch (e: NumberFormatException) {
+                                    // Invalid surface tension input
+                                    Toast.makeText(this, "Invalid surface tension input", Toast.LENGTH_SHORT).show()
+                                }
+                            }
+                            surfaceTensionInputDialog.show()
+                        }
+                    } catch (e: NumberFormatException) {
+                        // Invalid density input
+                        Toast.makeText(this, "Invalid density input", Toast.LENGTH_SHORT).show()
+                    }
+                }
+                densityInputDialog.show()
+
+            }
+            else{
+                Toast.makeText(this, "Invalid name input", Toast.LENGTH_SHORT).show()
+
+            }
+
+        }
+        nameInputDialog.show()
+
+    }
+
+    fun displayCustomLiquids() {
+        setContentView(R.layout.horizontal_scroll_liquids)
+
+        val liquidManager = LiquidManager()
+
+        val liquids = liquidManager.loadLiquids(getExternalFilesDir(Environment.DIRECTORY_PICTURES))
+
+        val goBackButton = findViewById<ImageButton>(R.id.buttonback123)
+        val addNewLiquidButton = findViewById<ImageButton>(R.id.buttonhej)
+
+        addNewLiquidButton.setOnClickListener{
+            customliquids()
+        }
+
+        val view = findViewById<LinearLayout>(R.id.linearlayout1)
+
+        //iterates through all the liquids and adds a button for choosing each liquid.
+        for (liquid in liquids) {
+            val btn = ImageButton(this)
+            when (liquid.name) {
+                "Blood" -> btn.setImageResource(R.drawable.blood)
+                "Water" -> btn.setImageResource(R.drawable.water)
+                "Oil" -> btn.setImageResource(R.drawable.oil)
+                else -> btn.setImageResource(R.drawable.custom_add_button)
+            }
+            btn.setOnClickListener {
+                displayIndividualCustom(liquid)
+            }
+            view.addView(btn)
+        }
+
+        goBackButton.setOnClickListener{
+            displaySettingsPage()
+        }
+
+
+    }
+
+    fun displayIndividualCustom(liquid: LiquidManager.Liquid){
+        setContentView(R.layout.display_individual_custom)
+
+        val confirmliquidbutton = findViewById<Button>(R.id.conliq)
+        val deletliquidbutton = findViewById<Button>(R.id.delustomC1)
+        val backbutton = findViewById<ImageButton>(R.id.goback)
+
+        val nametext = findViewById<TextView>(R.id.textViewname)
+        val dentext = findViewById<TextView>(R.id.textViewdensity)
+        val surftext = findViewById<TextView>(R.id.textViewsurfacetension)
+
+        nametext.text = liquid.name
+        dentext.text = liquid.density.toString()
+        surftext.text = liquid.surfaceTension.toString()
+
+        confirmliquidbutton.setOnClickListener{
+            currentLiquid = liquid
+            finish() // go back to landing page
+        }
+
+        deletliquidbutton.setOnClickListener{
+            if(liquid.name != "Blood" && liquid.name != "Water") {
+                val liquidManager = LiquidManager()
+                liquidManager.removeLiquid(
+                    liquid,
+                    getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+                )
+                displayCustomLiquids()
+            }
+        }
+
+        backbutton.setOnClickListener{
+
+            displayCustomLiquids()
+
+        }
+
+
+    }
+
+    fun displayAboutUsPage(){
+
+        setContentView(R.layout.aboutus)
+        val go_back = findViewById<ImageButton>(R.id.imageButton)
+        go_back.setOnClickListener{
+            displaySettingsPage()
+        }
+
+        val website = findViewById<ImageButton>(R.id.imageButton3)
+        website.setOnClickListener{
+            val url = "https://spilledowner.wixsite.com/spilled"
+            val intent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
+            startActivity(intent)
+        }
+    }
+
+}
