+++ app/src/main/res/layout/live_camera_page.xml
